// packages/mpl-core-kit-lib/src/generated/types/plugin.ts
export type Plugin =
  | { __kind: 'AddBlocker'; fields: readonly [AddBlocker] }
  | { __kind: 'Attributes'; fields: readonly [Attributes] }
  | { __kind: 'Autograph'; fields: readonly [Autograph] }
  | { __kind: 'BubblegumV2'; fields: readonly [BubblegumV2] }
  | { __kind: 'BurnDelegate'; fields: readonly [BurnDelegate] }
  | { __kind: 'Edition'; fields: readonly [Edition] }
  | { __kind: 'FreezeDelegate'; fields: readonly [FreezeDelegate] }
  | { __kind: 'FreezeExecute'; fields: readonly [FreezeExecute] }
  | { __kind: 'ImmutableMetadata'; fields: readonly [ImmutableMetadata] }
  | { __kind: 'MasterEdition'; fields: readonly [BaseMasterEdition] }
  | {
      __kind: 'PermanentBurnDelegate'
      fields: readonly [PermanentBurnDelegate]
    }
  | {
      __kind: 'PermanentFreezeDelegate'
      fields: readonly [PermanentFreezeDelegate]
    }
  | {
      __kind: 'PermanentFreezeExecute'
      fields: readonly [PermanentFreezeExecute]
    }
  | {
      __kind: 'PermanentTransferDelegate'
      fields: readonly [PermanentTransferDelegate]
    }
  | { __kind: 'Royalties'; fields: readonly [BaseRoyalties] }
  | { __kind: 'TransferDelegate'; fields: readonly [TransferDelegate] }
  | { __kind: 'UpdateDelegate'; fields: readonly [UpdateDelegate] }
  | { __kind: 'VerifiedCreators'; fields: readonly [VerifiedCreators] }