// packages/mpl-core-kit-lib/src/generated/types/pluginAuthorityPair.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  type Codec,
  combineCodec,
  type Decoder,
  type Encoder,
  getOptionDecoder,
  getOptionEncoder,
  getStructDecoder,
  getStructEncoder,
  type Option,
  type OptionOrNullable,
} from '@solana/kit'

import {
  type BasePluginAuthority,
  type BasePluginAuthorityArgs,
  getBasePluginAuthorityDecoder,
  getBasePluginAuthorityEncoder,
  getPluginDecoder,
  getPluginEncoder,
  type Plugin,
  type PluginArgs,
} from './index.ts'

export type PluginAuthorityPair = {
  authority: Option<BasePluginAuthority>
  plugin: Plugin
}

export type PluginAuthorityPairArgs = {
  authority: OptionOrNullable<BasePluginAuthorityArgs>
  plugin: PluginArgs
}

export function getPluginAuthorityPairCodec(): Codec<PluginAuthorityPairArgs, PluginAuthorityPair> {
  return combineCodec(getPluginAuthorityPairEncoder(), getPluginAuthorityPairDecoder())
}

export function getPluginAuthorityPairDecoder(): Decoder<PluginAuthorityPair> {
  return getStructDecoder([
    ['plugin', getPluginDecoder()],
    ['authority', getOptionDecoder(getBasePluginAuthorityDecoder())],
  ])
}

export function getPluginAuthorityPairEncoder(): Encoder<PluginAuthorityPairArgs> {
  return getStructEncoder([
    ['plugin', getPluginEncoder()],
    ['authority', getOptionEncoder(getBasePluginAuthorityEncoder())],
  ])
}