// packages/mpl-core-kit-lib/src/generated/types/registryRecord.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  type Codec,
  combineCodec,
  type Decoder,
  type Encoder,
  getStructDecoder,
  getStructEncoder,
  getU64Decoder,
  getU64Encoder,
} from '@solana/kit'

import {
  type BasePluginAuthority,
  type BasePluginAuthorityArgs,
  getBasePluginAuthorityDecoder,
  getBasePluginAuthorityEncoder,
  getPluginTypeDecoder,
  getPluginTypeEncoder,
  type PluginType,
  type PluginTypeArgs,
} from './index.ts'

export type RegistryRecord = {
  authority: BasePluginAuthority
  offset: bigint
  pluginType: PluginType
}

export type RegistryRecordArgs = {
  authority: BasePluginAuthorityArgs
  offset: bigint | number
  pluginType: PluginTypeArgs
}

export function getRegistryRecordCodec(): Codec<RegistryRecordArgs, RegistryRecord> {
  return combineCodec(getRegistryRecordEncoder(), getRegistryRecordDecoder())
}

export function getRegistryRecordDecoder(): Decoder<RegistryRecord> {
  return getStructDecoder([
    ['pluginType', getPluginTypeDecoder()],
    ['authority', getBasePluginAuthorityDecoder()],
    ['offset', getU64Decoder()],
  ])
}

export function getRegistryRecordEncoder(): Encoder<RegistryRecordArgs> {
  return getStructEncoder([
    ['pluginType', getPluginTypeEncoder()],
    ['authority', getBasePluginAuthorityEncoder()],
    ['offset', getU64Encoder()],
  ])
}